PN-2160: return empty result for bad endDate search (#191)
* PN-2160: return empty result for bad endDate search

* PN-2160: rm some code smells

+++ src/main/java/it/pagopa/pn/delivery/rest/PnReceivedNotificationsController.java
@@ -119,7 +119,7 @@ public class PnReceivedNotificationsController implements RecipientReadApi {
                     xPagopaPnCxType.toString(),
                     xPagopaPnCxId,
                     mandateId,
-                    docIdx.intValue()
+                    docIdx
             );
             logEvent.generateSuccess().log();
         } catch (Exception exc) {
